Multi-threaded array version, with Arc.
We need to pass the same array to everyone, but we need to move
it, so they don't know when to drop the underlying data. Enter
Arc which allows us to have ref-counted data which will be
garbage collected when the last reference is dropped. Arc::clone
clones the Arc, but not the underlying data.

This allows us to pass the array to the thread, but we cannot
update the value, because does not implement DerefMut; we can
look at the data, but we cannot mutate it.